+++ app/src/main/java/org/oxycblt/auxio/detail/recycler/AlbumDetailAdapter.kt
@@ -31,7 +31,6 @@ import org.oxycblt.auxio.ui.BindingViewHolder
 import org.oxycblt.auxio.ui.Item
 import org.oxycblt.auxio.ui.MenuItemListener
 import org.oxycblt.auxio.ui.SimpleItemCallback
-import org.oxycblt.auxio.ui.StyledImageView
 import org.oxycblt.auxio.util.context
 import org.oxycblt.auxio.util.formatDuration
 import org.oxycblt.auxio.util.getPluralSafe
@@ -194,21 +193,12 @@ private class AlbumSongViewHolder private constructor(private val binding: ItemA
                 isInvisible = false
                 contentDescription = context.getString(R.string.desc_track_number, item.track)
             }
-
-            binding.songTrackBg.setImageDrawable(null)
         } else {
             binding.songTrack.apply {
                 textSafe = ""
                 isInvisible = true
                 contentDescription = context.getString(R.string.def_track)
             }
-
-            // Normally we would not re-load the drawable each time and instead
-            // change the alpha value, but Lollipop gets in our way yet again
-            // and does some stupid insanity with the alpha value that results
-            // in such branching.
-            binding.songTrackBg.setImageDrawable(
-                StyledImageView.StyledDrawable(binding.context, R.drawable.ic_song))
         }
 
         binding.songName.textSafe = item.resolveName(binding.context)